commit before put to employees

+++ backend/server/routes/employees.py
@@ -2,7 +2,8 @@ from fastapi import APIRouter, HTTPException
 from server.models.employee import Employee
 from server.controllers.employees import (index_employee, list_employees, 
                                           create_employee, replace_employee,
-                                          destroy_employee)
+                                          destroy_employee, assign_employee_to_department,
+                                          delete_employee_department_relation)
 
 employee = APIRouter()
 
@@ -24,6 +25,8 @@ async def get_employees():
 async def post_employee(employee: Employee):
     response = await create_employee(employee)
     if response:
+        deptno = response.get("DEPTNO")
+        await assign_employee_to_department(employee.EMPNO, deptno)
         return response
     raise HTTPException(400, "Something went wrong")
 
@@ -32,7 +35,11 @@ async def put_employee(empno:int, data):#TODO tipe of data
     pass
 
 @employee.delete('/employees/{empno}')
-async def get_employee(empno:int):
+async def delete_employee(empno:int):
+    relationship_deleted = await delete_employee_department_relation(empno)
+    if not relationship_deleted:
+        raise HTTPException(404, f"No relationship found for employee with EMPNO {empno}")
+
     response = await destroy_employee(empno)
     if response:
         return "Successfully deleted employee"
